Added HazelcastStarterConstructor annotation to HazelcastStarter
This annotation marks ConstructorFunction classes for HazelcastStarter.
They are registered via reflection, so we don't have to adjust the
HazelcastProxyFactory each time, we add a new constructor class. This
also gives us support for EE only constructor classes, which cannot be
tested properly in OS.
