+++ hazelcast/src/test/java/com/hazelcast/test/starter/HazelcastProxyFactory.java
@@ -17,14 +17,7 @@
 package com.hazelcast.test.starter;
 
 import com.hazelcast.core.IFunction;
-import com.hazelcast.test.starter.constructor.AddressConstructor;
-import com.hazelcast.test.starter.constructor.ConfigConstructor;
-import com.hazelcast.test.starter.constructor.DataAwareEntryEventConstructor;
 import com.hazelcast.test.starter.constructor.EnumConstructor;
-import com.hazelcast.test.starter.constructor.GreaterLessPredicateConstructor;
-import com.hazelcast.test.starter.constructor.LifecycleEventConstructor;
-import com.hazelcast.test.starter.constructor.MapEventConstructor;
-import com.hazelcast.test.starter.constructor.VersionConstructor;
 import com.hazelcast.util.ConcurrentReferenceHashMap;
 import com.hazelcast.util.ConstructorFunction;
 import net.bytebuddy.ByteBuddy;
@@ -39,7 +32,9 @@ import net.bytebuddy.implementation.SuperMethodCall;
 import net.bytebuddy.implementation.attribute.MethodAttributeAppender;
 import net.bytebuddy.matcher.ElementMatchers;
 import net.bytebuddy.matcher.LatentMatcher;
+import org.reflections.Reflections;
 
+import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Proxy;
 import java.util.ArrayList;
@@ -47,19 +42,19 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedHashSet;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Queue;
 import java.util.Set;
-import java.util.concurrent.ConcurrentLinkedQueue;
 
 import static com.hazelcast.nio.ClassLoaderUtil.getAllInterfaces;
 import static com.hazelcast.test.starter.HazelcastAPIDelegatingClassloader.DELEGATION_WHITE_LIST;
 import static com.hazelcast.test.starter.HazelcastProxyFactory.ProxyPolicy.RETURN_SAME;
 import static com.hazelcast.test.starter.HazelcastStarterUtils.debug;
+import static com.hazelcast.test.starter.HazelcastStarterUtils.newCollectionFor;
+import static com.hazelcast.test.starter.ReflectionUtils.getConstructor;
+import static com.hazelcast.test.starter.ReflectionUtils.getReflectionsForTestPackage;
 import static com.hazelcast.util.ConcurrentReferenceHashMap.ReferenceType.STRONG;
+import static java.util.Arrays.asList;
 import static net.bytebuddy.jar.asm.Opcodes.ACC_PUBLIC;
 import static net.bytebuddy.matcher.ElementMatchers.is;
 import static net.bytebuddy.matcher.ElementMatchers.isConstructor;
@@ -68,7 +63,7 @@ public class HazelcastProxyFactory {
 
     // classes in this whitelist will not be proxied, instead instances of the same class (by name)
     // are constructed on target classloader
-    private static final Set<String> NO_PROXYING_WHITELIST;
+    private static final Map<String, Constructor<ConstructorFunction<Object, Object>>> NO_PROXYING_WHITELIST;
 
     // classes in this whitelist are explicitly selected for subclass proxying
     private static final Set<String> SUBCLASS_PROXYING_WHITELIST;
@@ -86,39 +81,46 @@ public class HazelcastProxyFactory {
     private static final ConcurrentReferenceHashMap<Class<?>, ConstructorFunction<Object, Object>> CONSTRUCTORS
             = new ConcurrentReferenceHashMap<Class<?>, ConstructorFunction<Object, Object>>(16, STRONG, STRONG);
 
-    private static final String CLASS_NAME_ENTRY_EVENT = "com.hazelcast.core.EntryEvent";
-    private static final String CLASS_NAME_LIFECYCLE_EVENT = "com.hazelcast.core.LifecycleEvent";
-    private static final String CLASS_NAME_DATA_AWARE_ENTRY_EVENT = "com.hazelcast.map.impl.DataAwareEntryEvent";
-    private static final String CLASS_NAME_MAP_EVENT = "com.hazelcast.core.MapEvent";
-    private static final String CLASS_NAME_CONFIG = "com.hazelcast.config.Config";
-    private static final String CLASS_NAME_CLIENT_CONFIG = "com.hazelcast.client.config.ClientConfig";
-    private static final String CLASS_NAME_ADDRESS = "com.hazelcast.nio.Address";
-    private static final String CLASS_NAME_VERSION = "com.hazelcast.version.Version";
-    private static final String CLASS_NAME_GREATER_LESS_PREDICATE = "com.hazelcast.query.impl.predicates.GreaterLessPredicate";
-
     static {
-        Set<String> notProxiedClasses = new HashSet<String>();
-        notProxiedClasses.add(CLASS_NAME_DATA_AWARE_ENTRY_EVENT);
-        notProxiedClasses.add(CLASS_NAME_MAP_EVENT);
-        notProxiedClasses.add(CLASS_NAME_CONFIG);
-        notProxiedClasses.add(CLASS_NAME_CLIENT_CONFIG);
-        notProxiedClasses.add(CLASS_NAME_ADDRESS);
-        notProxiedClasses.add(CLASS_NAME_VERSION);
-        notProxiedClasses.add(CLASS_NAME_GREATER_LESS_PREDICATE);
-        NO_PROXYING_WHITELIST = notProxiedClasses;
-
+        Map<String, Constructor<ConstructorFunction<Object, Object>>> notProxiedClasses
+                = new HashMap<String, Constructor<ConstructorFunction<Object, Object>>>();
         Set<String> subclassProxiedClasses = new HashSet<String>();
-        subclassProxiedClasses.add(CLASS_NAME_ENTRY_EVENT);
-        subclassProxiedClasses.add(CLASS_NAME_LIFECYCLE_EVENT);
-        SUBCLASS_PROXYING_WHITELIST = subclassProxiedClasses;
-
         Map<String, String> refactoredInterfaces = new HashMap<String, String>();
+
+        Reflections reflections = getReflectionsForTestPackage("com.hazelcast.test.starter.constructor");
+        Set<Class<?>> constructorClasses = reflections.getTypesAnnotatedWith(HazelcastStarterConstructor.class);
+        for (Class<?> constructorClass : constructorClasses) {
+            HazelcastStarterConstructor annotation = constructorClass.getAnnotation(HazelcastStarterConstructor.class);
+            String[] classNames = annotation.classNames();
+            switch (annotation.proxyPolicy()) {
+                case NO_PROXY:
+                    // we need to check that this is a valid ConstructorFunction class
+                    if (!ConstructorFunction.class.isAssignableFrom(constructorClass)) {
+                        throw new IllegalArgumentException("HazelcastStarterConstructor annotated class with ProxyPolicy.NO_PROXY"
+                                + " has to implement ConstructorFunction<Object, Object>: " + constructorClass);
+                    }
+                    // we lookup the constructor of the class and store it for each supported className
+                    Constructor<ConstructorFunction<Object, Object>> constructor = getConstructor(constructorClass, Class.class);
+                    for (String className : classNames) {
+                        notProxiedClasses.put(className, constructor);
+                    }
+                    break;
+                case SUBCLASS_PROXY:
+                    subclassProxiedClasses.addAll(asList(classNames));
+                    break;
+                default:
+                    throw new UnsupportedOperationException("Unsupported ProxyPolicy " + annotation.proxyPolicy());
+            }
+        }
+
+        NO_PROXYING_WHITELIST = notProxiedClasses;
+        SUBCLASS_PROXYING_WHITELIST = subclassProxiedClasses;
         REFACTORED_INTERFACES = refactoredInterfaces;
     }
 
     /**
-     * Decide whether given {@code delegateClass} should be proxied by subclassing, dynamic JDK proxy or not
-     * proxied at all.
+     * Decides whether the given {@code delegateClass} should be proxied by
+     * subclassing, dynamic JDK proxy or not proxied at all.
      *
      * @param delegateClass class of object to be proxied
      * @param ifaces        interfaces implemented by delegateClass
@@ -132,7 +134,7 @@ public class HazelcastProxyFactory {
         if (DELEGATION_WHITE_LIST.contains(className)) {
             return RETURN_SAME;
         }
-        if (NO_PROXYING_WHITELIST.contains(className) || delegateClass.isEnum()) {
+        if (NO_PROXYING_WHITELIST.containsKey(className) || delegateClass.isEnum()) {
             return ProxyPolicy.NO_PROXY;
         }
         if (SUBCLASS_PROXYING_WHITELIST.contains(className) || ifaces.length == 0) {
@@ -142,14 +144,16 @@ public class HazelcastProxyFactory {
     }
 
     /**
-     * This is the main entry point to obtain proxies for a target class loader.
-     * Create an Object valid for the Hazelcast version started with {@code targetClassLoader} that proxies
-     * the given {@code arg} which is valid in the current Hazelcast version.
+     * Main entry point to obtain proxies for a target class loader.
+     * <p>
+     * Creates an Object valid for the Hazelcast version started with
+     * {@code targetClassLoader} that proxies the given {@code arg},
+     * which is valid in the current Hazelcast version.
      */
     public static Object proxyObjectForStarter(ClassLoader targetClassLoader, Object arg) throws ClassNotFoundException {
-        // handle JDK collections (eg ArrayList etc)
+        // handle JDK collections (e.g. ArrayList)
         if (isJDKClass(arg.getClass()) && Collection.class.isAssignableFrom(arg.getClass())) {
-            Collection targetCollection = newCollectionFor(arg.getClass());
+            Collection<Object> targetCollection = newCollectionFor(arg.getClass());
             Collection collectionArg = (Collection) arg;
             for (Object o : collectionArg) {
                 targetCollection.add(proxyObjectForStarter(targetClassLoader, o));
@@ -186,8 +190,9 @@ public class HazelcastProxyFactory {
     }
 
     /**
-     * Convenience method to proxy an array of objects to be passed as arguments to a method on a class that is
-     * loaded by {@code targetClassLoader}
+     * Converts an array of arguments to the {@code targetClassLoader}, so they
+     * can be passed conveniently as an argument to a method of a class in the
+     * another classloader.
      */
     public static Object[] proxyArgumentsIfNeeded(Object[] args, ClassLoader targetClassLoader) throws ClassNotFoundException {
         if (args == null) {
@@ -207,21 +212,13 @@ public class HazelcastProxyFactory {
     }
 
     /**
-     * @return a new Collection object of a class that is assignable from the given type
+     * Generates a JDK dynamic proxy implementing the expected interfaces.
      */
-    static Collection newCollectionFor(Class type) {
-        if (Set.class.isAssignableFrom(type)) {
-            // original set might be ordered
-            return new LinkedHashSet();
-        } else if (List.class.isAssignableFrom(type)) {
-            return new ArrayList();
-        } else if (Queue.class.isAssignableFrom(type)) {
-            return new ConcurrentLinkedQueue();
-        } else if (Collection.class.isAssignableFrom(type)) {
-            return new LinkedList();
-        } else {
-            throw new UnsupportedOperationException("Cannot locate collection type for " + type);
-        }
+    @SuppressWarnings("unchecked")
+    private static <T> T generateProxyForInterface(Object delegate, ClassLoader proxyTargetClassloader,
+                                                   Class<?>... expectedInterfaces) {
+        InvocationHandler myInvocationHandler = new ProxyInvocationHandler(delegate);
+        return (T) Proxy.newProxyInstance(proxyTargetClassloader, expectedInterfaces, myInvocationHandler);
     }
 
     private static boolean isJDKClass(Class clazz) {
@@ -245,8 +242,8 @@ public class HazelcastProxyFactory {
 
     private static Object constructWithSubclassProxy(ClassLoader targetClassLoader, Object arg) throws ClassNotFoundException {
         // proxy class via subclassing the existing class implementation in the target targetClassLoader
-        Class<?> delegateClass = targetClassLoader.loadClass(arg.getClass().getName());
-        return proxyWithSubclass(targetClassLoader, arg, delegateClass);
+        targetClassLoader.loadClass(arg.getClass().getName());
+        return proxyWithSubclass(targetClassLoader, arg);
     }
 
     private static Object constructWithoutProxy(ClassLoader targetClassLoader, Object arg) throws ClassNotFoundException {
@@ -259,59 +256,40 @@ public class HazelcastProxyFactory {
         return construct(targetClass, arg);
     }
 
-    /**
-     * Generate a JDK dynamic proxy implementing the expected interfaces.
-     */
-    private static <T> T generateProxyForInterface(Object delegate, ClassLoader proxyTargetClassloader,
-                                                   Class<?>... expectedInterfaces) {
-        InvocationHandler myInvocationHandler = new ProxyInvocationHandler(delegate);
-        return (T) Proxy.newProxyInstance(proxyTargetClassloader, expectedInterfaces, myInvocationHandler);
-    }
-
-    private static Object proxyWithSubclass(ClassLoader targetClassLoader, final Object arg, Class<?> delegateClass) {
-        Class<?> targetClass;
-        ProxySource proxySource = ProxySource.of(arg.getClass(), targetClassLoader);
-        targetClass = PROXIES.applyIfAbsent(proxySource, new IFunction<ProxySource, Class<?>>() {
+    private static Object proxyWithSubclass(ClassLoader targetClassLoader, final Object delegate) {
+        ProxySource proxySource = ProxySource.of(delegate.getClass(), targetClassLoader);
+        Class<?> targetClass = PROXIES.applyIfAbsent(proxySource, new IFunction<ProxySource, Class<?>>() {
             @Override
             public Class<?> apply(ProxySource input) {
                 return new ByteBuddy().subclass(input.getToProxy(), AllAsPublicConstructorStrategy.INSTANCE)
                         .method(ElementMatchers.isDeclaredBy(input.getToProxy()))
-                        .intercept(InvocationHandlerAdapter.of(new ProxyInvocationHandler(arg)))
+                        .intercept(InvocationHandlerAdapter.of(new ProxyInvocationHandler(delegate)))
                         .make()
                         .load(input.getTargetClassLoader())
                         .getLoaded();
             }
         });
-        return construct(targetClass, arg);
+        return construct(targetClass, delegate);
     }
 
-    private static Object construct(Class<?> klass, Object delegate) {
-        ConstructorFunction<Object, Object> constructorFunction = CONSTRUCTORS.applyIfAbsent(klass,
+    private static Object construct(Class<?> clazz, Object delegate) {
+        ConstructorFunction<Object, Object> constructorFunction = CONSTRUCTORS.applyIfAbsent(clazz,
                 new IFunction<Class<?>, ConstructorFunction<Object, Object>>() {
                     @Override
                     public ConstructorFunction<Object, Object> apply(Class<?> input) {
                         String className = input.getName();
-                        if (className.equals(CLASS_NAME_DATA_AWARE_ENTRY_EVENT)) {
-                            return new DataAwareEntryEventConstructor(input);
-                        } else if (className.equals(CLASS_NAME_MAP_EVENT)) {
-                            return new MapEventConstructor(input);
-                        } else if (className.equals(CLASS_NAME_LIFECYCLE_EVENT)) {
-                            return new LifecycleEventConstructor(input);
-                        } else if (className.equals(CLASS_NAME_ADDRESS)) {
-                            return new AddressConstructor(input);
-                        } else if (className.equals(CLASS_NAME_CONFIG)
-                                || className.equals(CLASS_NAME_CLIENT_CONFIG)) {
-                            return new ConfigConstructor(input);
-                        } else if (className.equals(CLASS_NAME_VERSION)) {
-                            return new VersionConstructor(input);
+                        Constructor<ConstructorFunction<Object, Object>> constructor = NO_PROXYING_WHITELIST.get(className);
+                        if (constructor != null) {
+                            try {
+                                return constructor.newInstance(input);
+                            } catch (Exception e) {
+                                throw new IllegalStateException(e);
+                            }
                         } else if (input.isEnum()) {
                             return new EnumConstructor(input);
-                        } else if (className.equals(CLASS_NAME_GREATER_LESS_PREDICATE)) {
-                            return new GreaterLessPredicateConstructor(input);
-                        } else {
-                            throw new UnsupportedOperationException("Cannot construct target object for target class" + input
-                                    + " on classloader " + input.getClassLoader());
                         }
+                        throw new UnsupportedOperationException("Cannot construct target object for target " + input
+                                + " on classloader " + input.getClassLoader());
                     }
                 });
 
@@ -319,7 +297,8 @@ public class HazelcastProxyFactory {
     }
 
     /**
-     * Return all interfaces implemented by {@code type}, along with {@code type} itself if it is an interface
+     * Returns all interfaces implemented by {@code type}, along with
+     * {@code type} itself if it's an interface.
      */
     private static Class<?>[] getAllInterfacesIncludingSelf(Class<?> type) {
         Set<Class<?>> interfaces = new HashSet<Class<?>>(Arrays.asList(getAllInterfaces(type)));
@@ -336,8 +315,10 @@ public class HazelcastProxyFactory {
     }
 
     /**
-     * (Class toProxy, ClassLoader targetClassLoader) tuple that is used as a key for caching the generated
-     * proxy class for {@code toProxy} on {@code targetClassloader}.
+     * Tuple of {@link Class} {@code toProxy} and {@link ClassLoader}
+     * {@code targetClassloader} that is used as a key for caching the
+     * generated proxy class for {@code toProxy} on the
+     * {@code targetClassloader}.
      */
     private static class ProxySource {
