+++ hazelcast/src/test/java/com/hazelcast/test/starter/ReflectionUtils.java
@@ -16,20 +16,71 @@
 
 package com.hazelcast.test.starter;
 
+import org.reflections.Reflections;
+import org.reflections.scanners.MethodAnnotationsScanner;
+import org.reflections.scanners.SubTypesScanner;
+import org.reflections.scanners.TypeAnnotationsScanner;
+import org.reflections.util.ClasspathHelper;
+import org.reflections.util.ConfigurationBuilder;
+import org.reflections.util.FilterBuilder;
+
+import java.io.File;
+import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
+import java.util.Arrays;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 import static com.hazelcast.util.Preconditions.checkHasText;
 import static com.hazelcast.util.Preconditions.checkNotNull;
+import static java.net.URLClassLoader.newInstance;
 
 /**
  * Reflection utilities.
  */
+@SuppressWarnings("WeakerAccess")
 public final class ReflectionUtils {
 
     private ReflectionUtils() {
+    }
+
+    public static Reflections getReflectionsForTestPackage(String forPackage) {
+        try {
+            URL testClassesURL = new File("target/test-classes").toURI().toURL();
+            URLClassLoader classLoader = newInstance(new URL[]{testClassesURL}, ClasspathHelper.staticClassLoader());
+            return new Reflections(new ConfigurationBuilder()
+                    .addUrls(ClasspathHelper.forPackage(forPackage, classLoader))
+                    .addClassLoader(classLoader)
+                    .filterInputsBy(new FilterBuilder().includePackage(forPackage))
+                    .setScanners(
+                            new SubTypesScanner(false),
+                            new TypeAnnotationsScanner(),
+                            new MethodAnnotationsScanner()));
+        } catch (MalformedURLException e) {
+            throw new IllegalArgumentException(e);
+        }
+    }
+
+    @SuppressWarnings("unchecked")
+    public static <T> Constructor<T> getConstructor(Class<?> constructorClass, Class<?>... parameterTypes) {
+        try {
+            Constructor<T> constructor = (Constructor<T>) constructorClass.getConstructor(parameterTypes);
+            constructor.setAccessible(true);
+            return constructor;
+        } catch (NoSuchMethodException ignored) {
+            try {
+                Constructor<T> constructor = (Constructor<T>) constructorClass.getDeclaredConstructor(parameterTypes);
+                constructor.setAccessible(true);
+                return constructor;
+            } catch (NoSuchMethodException e) {
+                throw new IllegalArgumentException("Could not find constructor " + constructorClass.getSimpleName() + "("
+                        + Arrays.toString(parameterTypes) + ")", e);
+            }
+        }
     }
 
     public static Object getFieldValueReflectively(Object arg, String fieldName) throws IllegalAccessException {
